requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.0
python-multipart==0.0.9
anthropic==0.39.0
alembic==1.13.0
python-dotenv==1.0.1
pymupdf>=1.24.0
python-docx>=1.1.0
